Add ip[6]tables support for Pod Identity feature
Pod Identity introduced back at Dec 28, 2023
- https://aws.amazon.com/blogs/containers/amazon-eks-pod-identity-a-new-way-for-applications-on-eks-to-obtain-iam-credentials/

According to public documentation, Pod Identity uses the hostNetwork of the node and it uses port `80` and port `2703` on a link-local address on the node. This address is `169.254.170.23` for IPv4 and `[fd00:ec2::23]` for IPv6 clusters,
- https://docs.aws.amazon.com/eks/latest/userguide/pod-identities.html#pod-id-considerations
- https://docs.aws.amazon.com/eks/latest/userguide/pod-id-agent-setup.html

Adding it into default iptables/ip6tables would be required.